web/requirements.txt
adal
astroid
azure-common
azure-servicebus
azure-keyvault-secrets
azure-identity
certifi
cffi
chardet
click
colorama
cryptography
Flask
Flask-SQLAlchemy
idna
isodate
isort
itsdangerous
Jinja2
lazy-object-proxy
MarkupSafe
mccabe
msrest
msrestazure
oauthlib
psycopg2
pycparser
PyJWT
pylint
python-dateutil
python-http-client
requests
requests-oauthlib
sendgrid
six
SQLAlchemy
toml
uamqp
urllib3
Werkzeug==2.2.
wrapt
